Fix settings to follow smart/dumb components
